% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-pkg.R
\docType{package}
\name{RcppStudent-package}
\alias{RcppStudent}
\alias{RcppStudent-package}
\title{RcppStudent: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
RcppStudent is a package showing how to use Rcpp Modules to make available
a C++ class under S4.
}
