% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-exports.R
\name{Student}
\alias{Student}
\title{Create a Student Object from the Student C++ Class}
\arguments{
\item{name}{Name of Student}

\item{age}{Age of Student}

\item{male}{Is Student a Male?}
}
\value{
A \code{Student} object from the \emph{C++} Student Class.
}
\description{
Allows for the creation of a Student Object in \emph{C++} from \emph{R}
using the \emph{C++} Student class.
}
\examples{
##################
## Constructor

# Construct new Student object called "ben"
ben = new(Student, name = "Ben", age = 26, male = TRUE)

##################
## Getters

ben$LikesBlue()

ben$GetAge()

ben$IsMale()

ben$GetName()

ben$GetFavoriteNumbers()
}
